requirements.txt
Django==1.6.2
Pillow==2.3.1
South==0.8.4
django-tastypie==0.11.0
django-tastypie-swagger==0.1.2
gunicorn==18.0
numpy==1.9.0
paho-mqtt==0.4.94
psycopg2==2.5.2
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.6.1
wsgiref==0.1.2
xlwt==0.7.5
tornado==4.0.1
raven==5.0.0